AMQP-39 Add AMQP MessageBuilder
Fluent API for building messages.

Add equals() and toString() implementations for Message and MessageProperties.

JIRA: https://jira.springsource.org/browse/AMQP-39

AMQP-39 Polishing; PR Comments

Biggest change is to break out a separate MessagePropertiesBuilder.

However, the MessageBuilder uses one as a delegate so that the
fluent API can be used directly on the MessageBuilder without
having to build a separate MessageProperties first.

+++ spring-amqp/src/main/java/org/springframework/amqp/support/converter/Jackson2JsonMessageConverter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2013 the original author or authors. Licensed under the Apache License, Version 2.0 (the "License");
+ * Copyright 2002-2014 the original author or authors. Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy of the License at
  * http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
@@ -13,6 +13,7 @@ import java.io.IOException;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+
 import org.springframework.amqp.core.Message;
 import org.springframework.amqp.core.MessageProperties;
 
@@ -145,7 +146,7 @@ public class Jackson2JsonMessageConverter extends AbstractJsonMessageConverter {
 		messageProperties.setContentType(MessageProperties.CONTENT_TYPE_JSON);
 		messageProperties.setContentEncoding(getDefaultCharset());
 		if (bytes != null) {
-			messageProperties.setContentLength(bytes.length);
+			messageProperties.setContentLength((long) bytes.length);
 		}
 
 		if (getClassMapper() == null) {
